refactor(dprst_depth): promote spike topo functions to importable library
Move the 10 validated Phase 0 spike topo functions (dprst_polygons,
resolution_class, depth_to_spill, is_hydroflattened, volume_mean_depth,
lake_max_depth, max_to_mean, _tile13_name, _normalize_nodata, read_window)
plus their private helpers (_utm_zone_epsg, _1m_candidate_tiles,
_existing_paths, _native_resolution, _resolve_1m_paths) verbatim from
scripts/diagnose/dprst_depth_probe.py into src/gfv2_params/dprst_depth/topo.py,
so the upcoming Phase 1 builder can import them without duplicating logic.
The diagnostic now imports from the library; its CLI/analysis modes are
unchanged. Unit tests covering these functions move to
tests/test_dprst_depth_topo.py, importing directly from the package.
